+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Added the `clean` flag to **setup-env** to delete temp files that were created by `lint` from the repo.
 * Fixed an issue in **validate** where there is a specific value for marketplace in `isFetch` parameter.
 * Fixed an issue where the build number was incorrectly shown in XSOAR marketplace when using the **upload** command on external repositories via GitLab Pipelines.
+* Added support for the `List` content-item to **prepare-content** and **split**.
+* **Breaking change**: The **download** command will no longer download the content-item `List` in a unified file but instead will split the content-item `List` into two files (metadata and data).
 
 ## 1.25.0
 * Added support to detect automatically the playground-id when running cli commands in xsoar-6.
@@ -12,8 +14,6 @@
 * Fixed an issue where **validate** command failed on valid complex layout rules and triggers.
 * Fixed an issue where *validate* command failed with release notes files for assets modeling rules folder.
 * Added support for `skip` property in **pre-commit** hooks.
-* Added support of the `List` content-item to the **prepare-content** and **split** commands.
-* * **Breaking change**: The **download** command will no longer download the content-item `List` in a unified file but instead will split the content-item `List` into two files (metadata and data).
 * **generate-unit-tests** command will require installation with `pip install demisto-sdk[generate-unit-tests]`.
 * Added the *IN150* and *IN161* errors to *allowed ignore errors* list.
 * Added support for `env`, `copy_files` property in **pre-commit** docker hooks.